Game2048.java were split to two classes: root stack pane with keyboard and swipe events handlers and scene creation.
Now it can be shown in a WebFX tab without limitations (was used WSAD instead arrows workaround)
